Desacoplando e padronizando
