+++ app/routes/eventos.js
@@ -1,40 +1,7 @@
 const express = require('express');
 const router  = express.Router();
 const Evento  = require('../models/evento');
-const moment  = require('moment');
-const jwt     = require('jwt-simple');
-
-const middlewareAuth = function(req, res, next) {
-
-  const lancaErro = function(msg, code) {
-    let err = new Error(msg);
-    err.status = code;
-    return err;
-  }
-
-  let token = req.body.token || req.headers['x-access-token'];
-
-  try {
-
-    if (!token) {
-      throw lancaErro("Proibido", 403);
-    }
-
-    let decoded = jwt.decode(token, process.env.JWT_SECRET);
-    let expirou = moment(decoded.expira).isBefore(new Date());
-
-    if (expirou) {
-      throw lancaErro("Acesso não autorizado", 401);
-    }
-
-    req.user = decoded.username;
-    next();
-
-  } catch (e) {
-    return next(e);
-  }
-
-};
+const middlewareAuth = require('../utils/middlewareAuth');
 
 router.route('/')
   .post(middlewareAuth, function(req, res) {
